+++ src/fast/mod.rs
@@ -1,0 +1,5 @@
+mod address;
+mod fld;
+mod hasher;
+
+pub use hasher::FastStableHasher;
