+++ src/lib.rs
@@ -1,15 +1,43 @@
-pub(crate) mod crypto;
-mod field_address;
+pub mod crypto;
+pub mod fast;
 mod impls;
-mod mixers;
 pub mod prelude;
-mod stable_hash;
 // TODO: Move some things out of utils
 pub mod utils;
 
-pub use crate::crypto::set_hasher::SetHasher;
-pub use crate::field_address::FieldAddress;
-pub use crate::stable_hash::{StableHash, StableHasher};
-// TODO: Rename
-// TODO: Add from_bytes, to_bytes
-pub use crate::utils::StableHasherWrapper;
+use prelude::*;
+
+/// Like Hasher, but consistent across:
+/// * builds (independent of rustc version or std implementation details)
+/// * platforms (eg: 32 bit & 64 bit, x68 and ARM)
+/// * processes (multiple runs of the same program)
+pub trait StableHasher {
+    type Out: StableHash;
+    type Addr: FieldAddress;
+    fn write(&mut self, field_address: Self::Addr, bytes: &[u8]);
+    fn finish(&self) -> Self::Out;
+    fn new() -> Self;
+}
+
+/// Like Hash, but consistent across:
+/// * builds (independent of rustc version or std implementation details)
+/// * platforms (eg: 32 bit & 64 bit, x68 and ARM)
+/// * processes (multiple runs of the same program)
+pub trait StableHash {
+    fn stable_hash<H: StableHasher>(&self, field_address: H::Addr, state: &mut H);
+}
+
+pub trait FieldAddress: Clone {
+    fn root() -> Self;
+    fn child(&mut self, number: u64) -> Self;
+}
+
+pub fn fast_stable_hash<T: StableHash>(value: &T) -> u128 {
+    profile_fn!(stable_hash);
+    stable_hash_generic::<T, crate::fast::FastStableHasher>(value)
+}
+
+pub fn crypto_stable_hash<T: StableHash>(value: &T) -> [u8; 32] {
+    profile_fn!(crypto_stable_hash);
+    stable_hash_generic::<T, crate::crypto::CryptoStableHasher>(value)
+}
