+++ src/prelude.rs
@@ -1,6 +1,5 @@
 pub(crate) use crate::utils::*;
 pub(crate) use firestorm::{profile_fn, profile_method};
 
-pub use crate::field_address::FieldAddress;
+pub use crate::FieldAddress;
 pub use crate::{StableHash, StableHasher};
-// TODO: Add imports for the stable_hash and crypto_stable_hash
